+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Version 123:
 * Use unit-test subtree
 * Fix spurious race in websocket close test
 * Check compiler feature in Jamfile
+* Clear previous message fields in parser
 
 --------------------------------------------------------------------------------